// src/Web/ClientApp/src/app/home/programs/program-apply/achievements/achievements.interface.ts
export interface IAchievement {
  id?: string
  completed?: boolean
  description?: string
  description_AR?: string
  summaryOfAchievement?: string
  summaryOfAchievement_AR?: string
  populationImpact?: number
  achievementImpact?: number
  financialImpact?: number
  organization?: string
  organization_AR?: string
  yearOfAchievement?: string
}